Use enums rather than basic types for compact unwind row
Or, use the type system :)

Test Plan
=========

CI + some simple benchmarks with `--release` to ensure that performance
did not degrade

```
$ hyperfine "target/lightswitch-new --show-info /proc
/2154884/root/nix/store/dwzq3jgydz8wwa7xh86f8yfxg4isnjkj-clickhouse-24.3.2.23/bin/clickhouse" "target/light
switch-main --show-info /proc/2154884/root/nix/store/dwzq3jgydz8wwa7xh86f8yfxg4isnjkj-clickhouse-24.3.2.23/
bin/clickhouse"
Benchmark 1: target/lightswitch-new --show-info /proc/2154884/root/nix/store/dwzq3jgydz8wwa7xh86f8yfxg4isnjkj-clickhouse-24.3.2.23/bin/clickhouse
  Time (mean ± σ):     805.8 ms ±   9.5 ms    [User: 778.6 ms, System: 24.9 ms]
  Range (min … max):   792.8 ms … 817.7 ms    10 runs

Benchmark 2: target/lightswitch-main --show-info /proc/2154884/root/nix/store/dwzq3jgydz8wwa7xh86f8yfxg4isnjkj-clickhouse-24.3.2.23/bin/clickhouse
  Time (mean ± σ):     804.6 ms ±   9.9 ms    [User: 777.9 ms, System: 24.8 ms]
  Range (min … max):   792.7 ms … 825.0 ms    10 runs

Summary
  target/lightswitch-main --show-info /proc/2154884/root/nix/store/dwzq3jgydz8wwa7xh86f8yfxg4isnjkj-clickhouse-24.3.2.23/bin/clickhouse ran
    1.00 ± 0.02 times faster than target/lightswitch-new --show-info /proc/2154884/root/nix/store/dwzq3jgydz8wwa7xh86f8yfxg4isnjkj-clickhouse-24.3.2.23/bin/clickhouse
```

+++ src/unwind_info/convert.rs
@@ -138,11 +138,11 @@ impl<'a> CompactUnwindInfoBuilder<'a> {
                         match row.cfa() {
                             CfaRule::RegisterAndOffset { register, offset } => {
                                 if register == &RBP_X86 {
-                                    compact_row.cfa_type = CfaType::FramePointerOffset as u8;
+                                    compact_row.cfa_type = CfaType::FramePointerOffset;
                                 } else if register == &RSP_X86 {
-                                    compact_row.cfa_type = CfaType::StackPointerOffset as u8;
+                                    compact_row.cfa_type = CfaType::StackPointerOffset;
                                 } else {
-                                    compact_row.cfa_type = CfaType::UnsupportedRegisterOffset as u8;
+                                    compact_row.cfa_type = CfaType::UnsupportedRegisterOffset;
                                 }
 
                                 match u16::try_from(*offset) {
@@ -150,7 +150,7 @@ impl<'a> CompactUnwindInfoBuilder<'a> {
                                         compact_row.cfa_offset = off;
                                     }
                                     Err(_) => {
-                                        compact_row.cfa_type = CfaType::OffsetDidNotFit as u8;
+                                        compact_row.cfa_type = CfaType::OffsetDidNotFit;
                                     }
                                 }
                             }
@@ -167,29 +167,29 @@ impl<'a> CompactUnwindInfoBuilder<'a> {
                                     compact_row.cfa_offset = PltType::Plt2 as u16;
                                 }
 
-                                compact_row.cfa_type = CfaType::Expression as u8;
+                                compact_row.cfa_type = CfaType::Expression;
                             }
                         };
 
                         match row.register(RBP_X86) {
                             gimli::RegisterRule::Undefined => {}
                             gimli::RegisterRule::Offset(offset) => {
-                                compact_row.rbp_type = RbpType::CfaOffset as u8;
+                                compact_row.rbp_type = RbpType::CfaOffset;
 
                                 match i16::try_from(offset) {
                                     Ok(off) => {
                                         compact_row.rbp_offset = off;
                                     }
                                     Err(_) => {
-                                        compact_row.rbp_type = RbpType::OffsetDidNotFit as u8;
+                                        compact_row.rbp_type = RbpType::OffsetDidNotFit;
                                     }
                                 }
                             }
                             gimli::RegisterRule::Register(_reg) => {
-                                compact_row.rbp_type = RbpType::Register as u8;
+                                compact_row.rbp_type = RbpType::Register;
                             }
                             gimli::RegisterRule::Expression(_) => {
-                                compact_row.rbp_type = RbpType::Expression as u8;
+                                compact_row.rbp_type = RbpType::Expression;
                             }
                             _ => {
                                 // print!(", rbp unsupported {:?}", rbp);
@@ -199,7 +199,7 @@ impl<'a> CompactUnwindInfoBuilder<'a> {
                         if row.register(fde.cie().return_address_register())
                             == gimli::RegisterRule::Undefined
                         {
-                            compact_row.rbp_type = RbpType::UndefinedReturnAddress as u8;
+                            compact_row.rbp_type = RbpType::UndefinedReturnAddress;
                         }
                     }
                     _ => continue,
